+++ src/app/login/page.tsx
@@ -93,19 +93,34 @@ export default function Login() {
                 </p>
               )}
             </div>
-            <div className="flex items-center justify-between bg-[#F1F3F7] dark:bg-[#1E1E1E] border-st-edge dark:border-st-edgeDark rounded-sm border-[1.8px] h-[40px] w-[20rem] pl-4 ">
-              <input
-                {...register("password", {
-                  required: true,
-                  min: 8,
-                  max: 25,
-                })}
-                style={{ borderColor: errors.password ? "#b92828" : "" }}
-                className="bg-[#F1F3F7] dark:bg-[#1E1E1E] placeholder:text-[14px] dark:placeholder:text-st-edgeDark placeholder:text-[#5D6675] w-[85%] focus:outline-none dark:focus:border-white focus:border-[#3D4450]"
-                type={showPassword ? "text" : "password"}
-                name="password"
-                placeholder="Enter your password"
-              />
+            <>
+              <div className="flex items-center justify-between bg-[#F1F3F7] dark:bg-[#1E1E1E] border-st-edge dark:border-st-edgeDark rounded-sm border-[1.8px] h-[40px] w-[20rem] pl-4 ">
+                <input
+                  {...register("password", {
+                    required: true,
+                    min: 8,
+                    max: 25,
+                  })}
+                  style={{ borderColor: errors.password ? "#b92828" : "" }}
+                  className="bg-[#F1F3F7] dark:bg-[#1E1E1E] placeholder:text-[14px] dark:placeholder:text-st-edgeDark placeholder:text-[#5D6675] w-[85%] focus:outline-none dark:focus:border-white focus:border-[#3D4450]"
+                  type={showPassword ? "text" : "password"}
+                  name="password"
+                  placeholder="Enter your password"
+                />
+                <div className="w-[10%]">
+                  {showPassword ? (
+                    <AiOutlineEyeInvisible
+                      className="cursor-pointer ease duration-500"
+                      onClick={() => setShowPassword(!showPassword)}
+                    />
+                  ) : (
+                    <AiOutlineEye
+                      className="cursor-pointer ease duration-500"
+                      onClick={() => setShowPassword(!showPassword)}
+                    />
+                  )}
+                </div>
+              </div>
               {errors.password && (
                 <p className="text-[#b92828] text-[12px] text-center">
                   Password must be at least 8 characters, can contain at least
@@ -117,20 +132,7 @@ export default function Login() {
                   {responseError}
                 </p>
               )}
-              <div className="w-[10%]">
-                {showPassword ? (
-                  <AiOutlineEyeInvisible
-                    className="cursor-pointer ease duration-500"
-                    onClick={() => setShowPassword(!showPassword)}
-                  />
-                ) : (
-                  <AiOutlineEye
-                    className="cursor-pointer ease duration-500"
-                    onClick={() => setShowPassword(!showPassword)}
-                  />
-                )}
-              </div>
-            </div>
+            </>
             <p className="my-2 text-st-edgeDark text-[11px] font-bold">
               Forgot your{" "}
               <Link
